Backend/data_processor_service/requirements.txt
Django==4.2.16
confluent-kafka==2.5.3
pandas==2.2.3
numpy==1.26.4
python-dateutil==2.9.0
pytz==2024.1